fixed directory and bugs
